ensure timestamp is an int

+++ data_generator/polygon_data_service.py
@@ -958,7 +958,7 @@ class PolygonDataService(BaseDataService):
                 limit=1
             )
             for q in quotes:
-                return q.bid_price, q.ask_price, q.participant_timestamp/1_000_000  # convert ns back to ms
+                return q.bid_price, q.ask_price, int(q.participant_timestamp/1_000_000)  # convert ns back to ms
         else:
             # crypto
             return 0, 0, 0
